requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-admin-thumbnails==0.2.6
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
